Upgrade more internals to size_t.  Version change to 2.0 because of A…

+++ ffs/ffs_internal.h
@@ -33,8 +33,8 @@ struct _FFSBuffer {
 
 typedef struct _internal_iovec {
     void *iov_base;
-    int iov_offset;
-    int iov_len;
+    size_t iov_offset;
+    size_t iov_len;
 } internal_iovec;
 
 struct _IOContextStruct {
